1.3.4

+++ plugin.video.fenlight/resources/lib/caches/base_cache.py
@@ -9,8 +9,6 @@ delete_file, get_property, set_property, clear_property = kodi_utils.delete_file
 notification, confirm_dialog, ok_dialog, open_file = kodi_utils.notification, kodi_utils.confirm_dialog, kodi_utils.ok_dialog, kodi_utils.open_file
 path_exists, list_dirs, progress_dialog, make_directory = kodi_utils.path_exists, kodi_utils.list_dirs, kodi_utils.progress_dialog, kodi_utils.make_directory
 databases_path = path_join(userdata_path, 'databases/')
-current_dbs = ('navigator.db', 'watched.db', 'favourites.db', 'traktcache.db', 'maincache.db', 'lists.db',
-				'discover.db', 'metacache.db', 'debridcache.db', 'external.db', 'settings.db')
 database_path_raw = path_join(userdata_path, 'databases')
 navigator_db = translatePath(path_join(database_path_raw, 'navigator.db'))
 watched_db = translatePath(path_join(database_path_raw, 'watched.db'))
@@ -23,10 +21,13 @@ metacache_db = translatePath(path_join(database_path_raw, 'metacache.db'))
 debridcache_db = translatePath(path_join(database_path_raw, 'debridcache.db'))
 external_db = translatePath(path_join(database_path_raw, 'external.db'))
 settings_db = translatePath(path_join(database_path_raw, 'settings.db'))
+resolved_db = translatePath(path_join(database_path_raw, 'resolved.db'))
 database_timeout = 20
+current_dbs = ('navigator.db', 'watched.db', 'favourites.db', 'traktcache.db', 'maincache.db', 'lists.db',
+				'discover.db', 'metacache.db', 'debridcache.db', 'external.db', 'settings.db', 'resolved.db')
 database_locations = {
 'navigator_db': navigator_db, 'watched_db': watched_db, 'favorites_db': favorites_db, 'settings_db': settings_db, 'trakt_db': trakt_db, 'maincache_db': maincache_db,
-'metacache_db': metacache_db, 'debridcache_db': debridcache_db, 'lists_db': lists_db, 'discover_db': discover_db, 'external_db': external_db
+'metacache_db': metacache_db, 'debridcache_db': debridcache_db, 'lists_db': lists_db, 'discover_db': discover_db, 'external_db': external_db, 'resolved_db': resolved_db
 		}
 integrity_check = {
 'settings_db': ('settings',),
@@ -39,9 +40,9 @@ integrity_check = {
 'lists_db': ('lists',),
 'discover_db': ('discover',),
 'debridcache_db': ('debrid_data',),
-'external_db': ('results_data',)
+'external_db': ('results_data',),
+'resolved_db': ('resolved',)
 		}
-
 table_creators = {
 'navigator_db': (
 'CREATE TABLE IF NOT EXISTS navigator (list_name text, list_type text, list_contents text, unique (list_name, list_type))',),
@@ -78,7 +79,10 @@ last_played text, resume_id integer, title text, unique (db_type, media_id, seas
 'CREATE TABLE IF NOT EXISTS results_data (provider text not null, db_type text not null, tmdb_id text not null, title text, year integer, season text, episode text, results text, \
 expires integer, unique (provider, db_type, tmdb_id, title, year, season, episode))',),
 'discover_db': (
-'CREATE TABLE IF NOT EXISTS discover (id text not null unique, db_type text not null, data text)',)
+'CREATE TABLE IF NOT EXISTS discover (id text not null unique, db_type text not null, data text)',),
+'resolved_db': (
+'CREATE TABLE IF NOT EXISTS resolved (media_type text not null, tmdb_id text not null, provider text not null, name text not null, id integer not null, data text not null, \
+unique (media_type, tmdb_id))',)
 		}
 media_prop = 'fenlight.%s'
 BASE_GET = 'SELECT expires, data FROM %s WHERE id = ?'
@@ -150,21 +154,21 @@ def clear_cache(cache_type, silent=False):
 	def _confirm(): return silent or confirm_dialog()
 	success = True
 	if cache_type == 'meta':
-		from caches.trakt_cache import clear_trakt_movie_sets
 		from caches.meta_cache import delete_meta_cache
-		clear_trakt_movie_sets()
 		success = delete_meta_cache(silent=silent)
 	elif cache_type == 'internal_scrapers':
 		if not _confirm(): return
 		from apis import easynews_api
-		easynews_api.clear_media_results_database()
-		for item in ('pm_cloud', 'rd_cloud', 'ad_cloud', 'folders'): clear_cache(item, silent=True)
+		results = []
+		results.append(easynews_api.clear_media_results_database())
+		for item in ('pm_cloud', 'rd_cloud', 'ad_cloud', 'folders'): results.append(clear_cache(item, silent=True))
+		success = False not in results
 	elif cache_type == 'external_scrapers':
 		from caches.external_cache import external_cache
 		from caches.debrid_cache import debrid_cache
-		data = external_cache.delete_cache(silent=silent)
-		clear_debrid_result = debrid_cache.clear_cache()
-		success = (data, clear_debrid_result) == ('success', 'success')
+		results = []
+		for item in (external_cache, debrid_cache): results.append(item.clear_cache())
+		success = False not in results
 	elif cache_type == 'trakt':
 		from caches.trakt_cache import clear_all_trakt_cache_data
 		success = clear_all_trakt_cache_data(silent=silent)
@@ -185,17 +189,23 @@ def clear_cache(cache_type, silent=False):
 		from apis.alldebrid_api import AllDebridAPI
 		success = AllDebridAPI().clear_cache()
 	elif cache_type == 'folders':
+		if not _confirm(): return
 		from caches.main_cache import main_cache
-		main_cache.delete_all_folderscrapers()
+		success = main_cache.delete_all_folderscrapers()
 	elif cache_type == 'list':
 		if not _confirm(): return
 		from caches.lists_cache import lists_cache
-		lists_cache.delete_all_lists()
+		success = lists_cache.delete_all_lists()
+	elif cache_type == 'resolved':
+		if not _confirm(): return
+		from caches.resolved_cache import resolved_cache
+		success = resolved_cache.clear_cache()
 	else:# main
 		if not _confirm(): return
 		from caches.main_cache import main_cache
-		main_cache.delete_all()
+		success = main_cache.delete_all()
 	if not silent and success: notification('Success')
+	return success
 
 def clear_all_cache():
 	if not confirm_dialog(): return
@@ -203,7 +213,7 @@ def clear_all_cache():
 	line = 'Clearing....[CR]%s'
 	caches = (('meta', 'Meta Cache'), ('internal_scrapers', 'Internal Scrapers Cache'), ('external_scrapers', 'External Scrapers Cache'),
 			('trakt', 'Trakt Cache'), ('imdb', 'IMDb Cache'), ('list', 'List Data Cache', ), ('main', 'Main Cache', ),
-			('pm_cloud', 'Premiumize Cloud'), ('rd_cloud', 'Real Debrid Cloud'), ('ad_cloud', 'All Debrid Cloud'))
+			('pm_cloud', 'Premiumize Cloud'), ('rd_cloud', 'Real Debrid Cloud'), ('ad_cloud', 'All Debrid Cloud'), ('resolved', 'Resolved Sources'))
 	for count, cache_type in enumerate(caches, 1):
 		try:
 			progressDialog.update(line % (cache_type[1]), int(float(count) / float(len(caches)) * 100))
